requirement.txt
audioop-lts==0.2.2
av==15.1.0
bottle==0.13.4
bottle-websocket==0.2.9
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
coloredlogs==15.0.1
contourpy==1.3.3
ctranslate2==4.6.0
cycler==0.12.1
dlib==20.0.0
Eel==0.18.2
face-recognition==1.3.0
face_recognition_models==0.3.0
faster-whisper==1.2.0
filelock==3.13.1
flatbuffers==25.9.23
fonttools==4.60.0
fsspec==2024.6.1
future==1.0.0
fuzzywuzzy==0.18.0
gevent==25.9.1
gevent-websocket==0.10.1
greenlet==3.2.4
hf-xet==1.1.10
huggingface-hub==0.35.1
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.10
importlib_resources==6.5.2
jellyfish==1.2.0
Jinja2==3.1.4
joblib==1.5.2
kiwisolver==1.4.9
Levenshtein==0.27.1
llvmlite==0.45.0
MarkupSafe==3.0.2
matplotlib==3.10.6
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.3
noisereduce==3.0.3
numba==0.62.0
numpy==2.1.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
onnxruntime==1.23.0
openai-whisper==20250625
opencv-python==4.12.0.88
packaging==25.0
pillow==11.0.0
protobuf==6.32.1
PyAudio==0.2.14
pycparser==2.23
pydub==0.25.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-Levenshtein==0.27.1
pyttsx3==2.99
PyYAML==6.0.3
RapidFuzz==3.14.1
regex==2025.9.18
requests==2.32.5
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
scipy==1.16.2
semantic-version==2.10.0
sentencepiece==0.2.1
setuptools==80.9.0
setuptools-rust==1.12.0
six==1.17.0
sounddevice==0.5.2
soundfile==0.13.1
speechbrain==1.0.3
SpeechRecognition==3.14.3
srt==3.5.3
standard-aifc==3.13.0
standard-chunk==3.13.0
sympy==1.13.3
tiktoken==0.11.0
tokenizers==0.22.1
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.1
triton==3.4.0
typing_extensions==4.12.2
urllib3==2.5.0
vosk==0.3.45
webrtcvad-wheels==2.0.14
websockets==15.0.1
zope.event==6.0
zope.interface==8.0.1
